safari-600.7-branch/LayoutTests/mathml/presentation/multiscript-superscriptshift.html

<!DOCTYPE html>
<html>
    <head>
        <script src="../../resources/js-test-pre.js"></script>
        <script>
            function runTest() {
                description('Tests that superscriptshift works on mmultiscripts. After adding superscriptshift the height of mmultiscripts element should increase');

                var rect = document.getElementById("multiscriptEle").getBoundingClientRect();
                height = rect.bottom - rect.top;

                shouldBeTrue("height > 150");
                isSuccessfullyParsed();
            }
        </script>
    </head>
    <body onload="runTest()">
        <math>
            <mmultiscripts superscriptshift="60px" id="multiscriptEle">
                <mspace width="50px" height="50px" mathbackground="green"/>
                <mspace width="50px" height="50px" mathbackground="blue"/>
                <mspace width="50px" height="50px" mathbackground="red"/>
				<mprescripts />
				<mspace width="50px" height="50px" mathbackground="yellow"/>
                <mspace width="50px" height="50px" mathbackground="black"/>
            </mmultiscripts>
      </math>
    </body>
</html>
