safari-600.7-branch/LayoutTests/mathml/presentation/multiscript-superscriptshift-expected.txt

Tests that superscriptshift works on mmultiscripts. After adding superscriptshift the height of mmultiscripts element should increase

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS height > 150 is true
PASS successfullyParsed is true

TEST COMPLETE
